Split test and updates to passport
