+++ expressjs/app/api/redacted/index.js
@@ -1,0 +1,3 @@
+module.exports = {
+  router: require('./router')
+}
