+++ expressjs/app/api/safepaths/index.js
@@ -1,0 +1,3 @@
+module.exports = {
+  router: require('./router')
+}
