+++ expressjs/test/lib/agent.js
@@ -1,0 +1,3 @@
+const { server } = require('src/')
+
+module.exports = server
